exercises 4.20-4.21

+++ part4/blog/controllers/blogsRouter.js
@@ -4,29 +4,28 @@ const Blog = require("../models/Blog");
 const User = require("../models/User")
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-      return authorization.substring(7)
-    }
-    return null
-  }
+
   
 
 
 blogsRouter.get("/", async (request, response) => {
+
     const blogs = await Blog.find({}).populate("user", {username: 1, name:1})
     response.json(blogs)
 });
 
 blogsRouter.post("/", async (request, response) => {
     const body = request.body;
-    const token = getTokenFrom(request);
+    const token = request.token;
     const decodedToken = jwt.verify(token, process.env.SECRET)
     if (!token || !decodedToken.id) {
       return response.status(401).json({ error: 'token missing or invalid' })
     }
+    console.log(
+        "Decodec", decodedToken
+    )
     const user = await User.findById(decodedToken.id)
+    console.log("user", user)
     try{
         const blog = new Blog({
             title: body.title,
@@ -42,7 +41,7 @@ blogsRouter.post("/", async (request, response) => {
         response.status(201).json(data);
     }catch(e) {
         console.log(e.name === 'ValidationError')
-        console.log(e)
+        //console.log(e)
         response.status(400).end()
     }
 })
@@ -60,11 +59,22 @@ blogsRouter.put("/:id", async (request, response, next) => {
 });
 
 blogsRouter.delete("/:id", async (request, response, next) => {
-    const data = await Blog.findByIdAndDelete(request.params.id)
-    try {
-        response.status(204).end()
-    } catch (e) {
-        next(e)
+    const token = request.token;
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    
+    const userId = decodedToken.id;
+    const targetedBlog = await Blog.findById(request.params.id);
+    if(targetedBlog.user.toString() == userId.toString()){
+        
+        const data = await Blog.findByIdAndDelete(request.params.id)
+        try {
+            response.status(204).end()
+        } catch (e) {
+            next(e)
+        }
     }
     
 })
